+++ src/form-validator/index.ts
@@ -1,7 +1,3 @@
-export { default as FormField } from './components/FormField.vue'
-export { default as FormValidatorProvider } from './components/FormValidatorProvider'
-export type { FormValidatorProviderExposed } from './components/FormValidatorProvider'
-
 export * from './FormValidator'
 export * from './injectionKey'
 export * from './types'