+++ src/image/index.ts
@@ -1,5 +1,2 @@
-export { default as Image } from './components/Image.astro'
-export { default as ResponsiveImage } from './components/ResponsiveImage.astro'
-
 export * from './fragments'
 export * from './types'