+++ src/page/index.ts
@@ -1,5 +1,3 @@
-export { default as PageFieldRender } from './components/PageFieldRender.astro'
-
 export * from './field'
 export * from './seo-meta'
 export * from './types'