+++ src/query-params/index.ts
@@ -1,5 +1,3 @@
-export { default as ProvideUrlConfig } from './components/ProvideUrlConfig.astro'
-
 export * from './config'
 export * from './store'
 export * from './types'